Refactor cpu
