+++ pkg/gohai/cpu/cpu.go
@@ -1,0 +1,14 @@
+package cpu
+
+type Cpu struct{}
+
+const name = "cpu"
+
+func (self *Cpu) Name() string {
+	return name
+}
+
+func (self *Cpu) Collect() (result interface{}, err error) {
+	result, err = getCpuInfo()
+	return
+}
