+++ pkg/gohai/cpu/cpu_linux.go
@@ -6,19 +6,6 @@ import (
 	"regexp"
 )
 
-type Cpu struct{}
-
-const name = "cpu"
-
-func (self *Cpu) Name() string {
-	return name
-}
-
-func (self *Cpu) Collect() (result interface{}, err error) {
-	result, err = getCpuInfo()
-	return
-}
-
 var cpuMap = map[string]string{
 	"vendor_id":  "vendor_id",
 	"model name": "model_name",
